requirements.txt
gradio_client
Pillow
